+++ frontend/src/components/TableView/GeneralTableView.vue
@@ -6,6 +6,13 @@
       border>
       <template v-for="col in columns">
         <el-table-column :key="col" :label="col" :property="col" min-width="120">
+          <template slot-scope="scope">
+            <el-popover trigger="hover" :content="getString(scope.row[col])" popper-class="cell-popover">
+              <div slot="reference" class="wrapper">
+                {{getString(scope.row[col])}}
+              </div>
+            </el-popover>
+          </template>
         </el-table-column>
       </template>
     </el-table>
@@ -58,23 +65,39 @@ export default {
   computed: {
     filteredData () {
       return this.data
-      // .map(d => d)
-      // .filter((d, index) => {
-      //   // pagination
-      //   const pageNum = this.pageNum
-      //   const pageSize = this.pageSize
-      //   return (pageSize * (pageNum - 1) <= index) && (index < pageSize * pageNum)
-      // })
     }
   },
   methods: {
     onPageChange () {
       this.$emit('page-change', { pageNum: this.pageNum, pageSize: this.pageSize })
+    },
+    getString (value) {
+      if (value === undefined) return ''
+      const str = JSON.stringify(value)
+      if (str.match(/^"(.*)"$/)) return str.match(/^"(.*)"$/)[1]
+      return str
     }
   }
 }
 </script>
 
 <style scoped>
+  .general-table-view >>> .cell .wrapper:hover {
+    text-decoration: underline;
+  }
 
+  .general-table-view >>> .cell .wrapper {
+    font-size: 12px;
+    height: 24px;
+    line-height: 24px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
+</style>
+
+<style>
+  .cell-popover {
+    max-width: 480px;
+  }
 </style>